api-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary

firebase-admin

google-cloud-tasks
httpx
